+++ Panther/src/sema/ScopeLevel.cpp
@@ -218,6 +218,9 @@ namespace pcit::panther::sema{
 	}
 
 
+	auto ScopeLevel::addIdent(std::string_view ident, sema::FuncAliasID id) -> AddIdentResult {
+		return this->add_ident_default_impl(ident, id);
+	}
 
 	auto ScopeLevel::addIdent(std::string_view ident, sema::VarID id) -> AddIdentResult {
 		return this->add_ident_default_impl(ident, id);